refactor ContentView
