+++ PlantApp/Views/Content View/ContentView.swift
@@ -1,0 +1,68 @@
+//
+//  ContentARView.swift
+//  PlantCodeiOS15
+//
+//  Created by Aeliana Shen on 6/5/23.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    @StateObject var viewModel = ContentViewModel()
+    @StateObject var cartManager = CartManager()
+    
+    var body: some View {
+        if viewModel.isOnboardingViewActive {
+            OnboardingView()
+        } else {
+            navigationViews
+        }
+    }
+    
+    private var navigationViews: some View {
+        NavigationView {
+            TabView {
+                ARDetectView()
+                    .tabItem {
+                        Image(systemName: "camera.viewfinder")
+                        Text("ARDetect")
+                    }
+                ShopView()
+                    .tabItem {
+                        Image(systemName: "leaf")
+                        Text("Plants")
+                    }
+                    .environmentObject(cartManager)
+                PreferSetView()
+                    .tabItem {
+                        Image(systemName: "heart")
+                        Text("Prefer")
+                    }
+                AccountView()
+                    .tabItem {
+                        Image(systemName: "person")
+                        Text("Account")
+                    }
+            }
+            .ignoresSafeArea(.keyboard)
+            .toolbar{
+                cartTool
+            }
+        }
+    }
+    
+    private var cartTool: some View {
+        NavigationLink {
+            CartView(viewModel: CartViewModel(cartManager: cartManager))
+                .environmentObject(cartManager)
+        } label: {
+            CartButton(viewModel: CartButtonViewModel(numOfProducts: cartManager.products.count))
+        }
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
